+++ packages/_example/src/forest/customizations/store.ts
@@ -4,7 +4,6 @@ import { Schema } from '../typings';
 export default (collection: Collection<Schema, 'store'>) =>
   collection
     .addManyToOneRelation('owner', 'owner', { foreignKey: 'ownerId' })
-    .addOneToOneRelation('address', 'location', { originKey: 'storeId' })
     .addOneToManyRelation('dvds', 'dvd', { originKey: 'storeId' })
     .addOneToManyRelation('accounts', 'account', { originKey: 'storeId' })