+++ packages/datasource-live/test/index.test.ts
@@ -1,9 +1,0 @@
-import { createLiveDataSource } from '../src';
-
-describe('exports', () => {
-  describe.each([['createLiveDataSource', createLiveDataSource]])('class %s', (message, type) => {
-    it('should be defined', () => {
-      expect(type).toBeDefined();
-    });
-  });
-});