+++ packages/datasource-sql/test/introspection/helpers/sql-type-converter.test.ts
@@ -1,6 +1,6 @@
 import { ColumnDescription, DataTypes, Sequelize } from 'sequelize';
 
-import SqlTypeConverter from '../../src/utils/sql-type-converter';
+import SqlTypeConverter from '../../../src/introspection/helpers/sql-type-converter';
 
 const makeColumnDescription = (description: Partial<ColumnDescription>) => {
   return {