+++ packages/datasource-toolkit/src/decorators/search/collection.ts
@@ -1,8 +1,9 @@
 import { validate as uuidValidate } from 'uuid';
 
 import { Caller } from '../../interfaces/caller';
-import { CollectionSchema, ColumnSchema, FieldSchema } from '../../interfaces/schema';
-import { DataSource } from '../../interfaces/collection';
+import { Collection } from '../../interfaces/collection';
+import { CollectionSchema, ColumnSchema } from '../../interfaces/schema';
+import { Operator } from '../../interfaces/query/condition-tree/nodes/operators';
 import { SearchReplacer } from './types';
 import CollectionCustomizationContext from '../../context/collection-context';
 import CollectionDecorator from '../collection-decorator';
@@ -27,7 +28,7 @@ export default class SearchCollectionDecorator extends CollectionDecorator {
     filter?: PaginatedFilter,
   ): Promise<PaginatedFilter> {
     // Search string is not significant
-    if (!filter?.search || SearchCollectionDecorator.checkEmptyString(filter.search)) {
+    if (!filter?.search?.trim().length) {
       return filter.override({ search: null });
     }
 
@@ -55,12 +56,7 @@ export default class SearchCollectionDecorator extends CollectionDecorator {
   }
 
   private defaultReplacer(search: string, extended: boolean): ConditionTree {
-    const searchableFields = SearchCollectionDecorator.getSearchFields(
-      this.childCollection.schema,
-      this.childCollection.dataSource,
-      extended,
-    );
-
+    const searchableFields = SearchCollectionDecorator.getFields(this.childCollection, extended);
     const conditions = searchableFields
       .map(([field, schema]) => SearchCollectionDecorator.buildCondition(field, schema, search))
       .filter(Boolean);
@@ -73,82 +69,63 @@ export default class SearchCollectionDecorator extends CollectionDecorator {
     schema: ColumnSchema,
     searchString: string,
   ): ConditionTree {
-    const { columnType, enumValues } = schema;
+    const { columnType, enumValues, filterOperators } = schema;
     const isNumber = Number(searchString).toString() === searchString;
     const isUuid = uuidValidate(searchString);
 
-    if (columnType === 'Number' && isNumber) {
+    if (columnType === 'Number' && isNumber && filterOperators?.has('Equal')) {
       return new ConditionTreeLeaf(field, 'Equal', Number(searchString));
     }
 
-    if (columnType === 'Enum') {
+    if (columnType === 'Enum' && filterOperators?.has('Equal')) {
       const searchValue = SearchCollectionDecorator.lenientFind(enumValues, searchString);
+
       if (searchValue) return new ConditionTreeLeaf(field, 'Equal', searchValue);
     }
 
     if (columnType === 'String') {
-      return new ConditionTreeLeaf(field, 'IContains', searchString);
+      const isCaseSensitive = searchString.toLocaleLowerCase() !== searchString.toLocaleUpperCase();
+      const supportsIContains = filterOperators?.has('IContains');
+      const supportsContains = filterOperators?.has('Contains');
+      const supportsEqual = filterOperators?.has('Equal');
+
+      // Perf: don't use case-insensitive operator when the search string is indifferent to case
+      let operator: Operator;
+      if (supportsIContains && (isCaseSensitive || !supportsContains)) operator = 'IContains';
+      else if (supportsContains) operator = 'Contains';
+      else if (supportsEqual) operator = 'Equal';
+
+      if (operator) return new ConditionTreeLeaf(field, operator, searchString);
     }
 
-    if (columnType === 'Uuid' && isUuid) {
+    if (columnType === 'Uuid' && isUuid && filterOperators?.has('Equal')) {
       return new ConditionTreeLeaf(field, 'Equal', searchString);
     }
 
     return null;
   }
 
-  private static lenientFind(enumValues: string[], searchString: string): string {
-    return (
-      enumValues?.find(v => v === searchString.trim()) ??
-      enumValues?.find(v => v.toLocaleLowerCase() === searchString.toLocaleLowerCase().trim())
-    );
-  }
+  private static getFields(collection: Collection, extended: boolean): [string, ColumnSchema][] {
+    const fields: [string, ColumnSchema][] = [];
 
-  private static getSearchFields(
-    schema: CollectionSchema,
-    dataSource: DataSource,
-    searchExtended: boolean,
-  ): [string, ColumnSchema][] {
-    const fields = Object.entries(schema.fields);
+    for (const [name, field] of Object.entries(collection.schema.fields)) {
+      if (field.type === 'Column') fields.push([name, field]);
 
-    if (searchExtended) SearchCollectionDecorator.getDeepFields(dataSource, fields);
+      if (extended && (field.type === 'ManyToOne' || field.type === 'OneToOne')) {
+        const related = collection.dataSource.getCollection(field.foreignCollection);
 
-    return fields.filter(([, fieldSchema]) =>
-      SearchCollectionDecorator.isSearchable(fieldSchema),
-    ) as [string, ColumnSchema][];
-  }
-
-  private static getDeepFields(dataSource: DataSource, fields: [string, FieldSchema][]) {
-    fields.forEach(([name, field]) => {
-      if (field.type === 'ManyToOne' || field.type === 'OneToOne') {
-        const related = dataSource.getCollection(field.foreignCollection);
-        fields.push(
-          ...Object.entries(related.schema.fields).map(
-            ([subName, columnSchema]) =>
-              [`${name}:${subName}`, columnSchema] as [string, FieldSchema],
-          ),
-        );
-      }
-    });
-  }
-
-  private static checkEmptyString(searchString: string) {
-    return searchString.trim().length === 0;
-  }
-
-  private static isSearchable(schema: FieldSchema): boolean {
-    if (schema.type === 'Column') {
-      const { columnType, filterOperators } = schema;
-
-      if (columnType === 'Enum' || columnType === 'Number' || columnType === 'Uuid') {
-        return filterOperators?.has('Equal');
-      }
-
-      if (columnType === 'String') {
-        return filterOperators?.has('IContains');
+        for (const [subName, subField] of Object.entries(related.schema.fields))
+          if (subField.type === 'Column') fields.push([`${name}:${subName}`, subField]);
       }
     }
 
-    return false;
+    return fields;
+  }
+
+  private static lenientFind(haystack: string[], needle: string): string {
+    return (
+      haystack?.find(v => v === needle.trim()) ??
+      haystack?.find(v => v.toLocaleLowerCase() === needle.toLocaleLowerCase().trim())
+    );
   }
 }