+++ packages/datasource-toolkit/src/index.ts
@@ -19,17 +19,18 @@ export { default as RenameCollectionDataSourceDecorator } from './decorators/ren
 export { default as ActionCollectionDecorator } from './decorators/actions/collection';
 export { default as ComputedCollectionDecorator } from './decorators/computed/collection';
 export { default as EmptyCollectionDecorator } from './decorators/empty/collection';
-export { default as RelationCollectionDecorator } from './decorators/relation/collection';
+export { default as HookCollectionDecorator } from './decorators/hook/collection';
 export { default as OperatorsEmulateCollectionDecorator } from './decorators/operators-emulate/collection';
 export { default as OperatorsReplaceCollectionDecorator } from './decorators/operators-replace/collection';
 export { default as PublicationCollectionDecorator } from './decorators/publication/collection';
+export { default as RelationCollectionDecorator } from './decorators/relation/collection';
 export { default as RenameFieldCollectionDecorator } from './decorators/rename-field/collection';
-export { default as SearchCollectionDecorator } from './decorators/search/collection';
-export { default as WriteCollectionDecorator } from './decorators/write/collection';
 export { default as SchemaCollectionDecorator } from './decorators/schema/collection';
+export { default as SearchCollectionDecorator } from './decorators/search/collection';
 export { default as SegmentCollectionDecorator } from './decorators/segment/collection';
 export { default as SortEmulateCollectionDecorator } from './decorators/sort-emulate/collection';
-export { default as HookCollectionDecorator } from './decorators/hook/collection';
+export { default as ValidationCollectionDecorator } from './decorators/validation/collection';
+export { default as WriteCollectionDecorator } from './decorators/write/collection';
 
 // Decorators (helpers)
 export { Action as ActionDefinition } from './decorators/actions/types/actions';